requirements.txt
ag2>=0.11.4,<0.12
openai>=2.29.0,<3
pydantic>=2.12.5,<3
python-dotenv>=1.2.2,<2
streamlit>=1.55.0,<2
fpdf2>=2.8.7,<3
